requirements.txt
Flask
Flask-SocketIO
gevent
gevent-websocket
PyYAML
SQLAlchemy
alembic
